requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
duckduckgo_search==7.2.0
e2b==1.0.5
e2b-code-interpreter==1.0.3
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gherkin-official==29.0.0
gradio==5.10.0
gradio_client==1.5.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.57.1
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.3
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parse_type==0.6.4
pillow==11.1.0
playwright==1.49.1
pluggy==1.5.0
primp==0.9.3
propcache==0.2.1
protobuf==5.29.2
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.4
pytest-bdd==8.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.6
safehttpx==0.1.6
safetensors==0.5.1
selenium==4.27.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smolagents==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
pytest-html==4.1.1